+++ src/main/java/com/sakadream/test/model/Database.java
@@ -33,8 +33,8 @@ public class Database {
             stmt = conn.createStatement();
         ResultSet rs = stmt.executeQuery("SELECT * FROM public.\"EMPLOYEES\"");
         while (rs.next()) {
-            Employee e = new Employee(rs.getLong(1), rs.getString(2), rs.getString(3), rs.getString(4), rs.getString(5),
-                    rs.getLong(6));
+            Employee e = new Employee(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getString(4), rs.getString(5),
+                    rs.getInt(6));
             list.add(e);
         }
         return list;
